# Speed up startup