--- a/nixui/options/api.py
+++ b/nixui/options/api.py
@@ -13,24 +13,19 @@ def get_option_tree(configuration_path=None):
     if configuration_path is None:
         configuration_path = os.environ['CONFIGURATION_PATH']
 
-    def convert_type_str_to_type_obj(option_data_dict, missing_default=None):
-        option_data_dict = copy.deepcopy(option_data_dict)
-        for key, value in option_data_dict.items():
-            option_data_dict[key]['type_string'] = option_data_dict[key]['type']
-            if 'type' in value:
-                option_data_dict[key]['type'] = types.from_nix_type_str(option_data_dict[key]['type'])
-            elif missing_default:
-                option_data_dict[key]['type'] = missing_default
-        return option_data_dict
+    system_option_data_dict = remap_dict.key_remapper(
+        nix_eval.get_all_nixos_options(),
+        {'system_default': 'system_default_definition'}
+    )
+    for key, value in system_option_data_dict.items():
+        system_option_data_dict[key]['type_string'] = system_option_data_dict[key]['type']
+        if 'type' in value:
+            system_option_data_dict[key]['type'] = types.from_nix_type_str(system_option_data_dict[key]['type'])
+        else:
+            system_option_data_dict[key]['type'] = types.AttrsType()
 
     return option_tree.OptionTree(
-        convert_type_str_to_type_obj(
-            remap_dict.key_remapper(
-                nix_eval.get_all_nixos_options(),
-                {'system_default': 'system_default_definition'}
-            ),
-            missing_default=types.AttrsType()
-        ),
+        system_option_data_dict,
         parser.get_all_option_values(configuration_path)
     )