Before: 55 seconds

After: 17 seconds

fixes https://github.com/nix-gui/nix-gui/issues/194

Changes to improve time:
- removed deepcopy from `api.get_option_tree` (55 seconds -> 53 seconds)
- load option types from option type strings at runtime rather than when starting the program (53 seconds -> 17 seconds)

Further work expected to reduce startup time to ~7 seconds https://github.com/nix-gui/nix-gui/issues/197 (outside of the scope of this PR)